refactor(engine): extract WeightUploadOrchestrator (init_weights) to its own TU
Moves Engine::init_weights (~230 LOC) from src/runtime/engine.cpp to
src/runtime/engine_weight_upload.cpp. Declaration in engine.h unchanged.
Body byte-identical.

Phase 4 of docs/superpowers/specs/2026-05-20-architecture-refactor-roadmap-design.md
