+++ CMakeLists.txt
@@ -221,6 +221,7 @@ set(IMP_RUNTIME_SOURCES
     src/runtime/batch.cpp
     src/runtime/engine.cpp
     src/runtime/engine_init_resolver.cpp
+    src/runtime/engine_weight_upload.cpp
     src/runtime/engine_graph_decode.cpp
     src/runtime/vision_pipeline.cpp
     src/runtime/constraint_manager.cpp
